:recycle:  move group handling to package utils
