+++ packages/core/src/index.js
@@ -2,6 +2,5 @@
 
 module.exports = {
   generateDocFromSchema: require("./generator").generateDocFromSchema,
-  groupInfo: require("./group-info"),
   config: require("./config"),
 };
